Added the initial service-level triton deployment package
To get this worked I needed the config.pbtxt files inputs/outputs
to match by name. It would be nice to be able to rename the output
of one model that needs to be fed into the next model to not need
the same names.

But for now we can live with this