Added docker-compose & reverting back to simple approach without batching
I'm going to make this a different branch. The idea is to have a different
branch for each layer of sophistication.

v1: Following the basic documentation where you call the model for each request
v2: Turn on dynamic batching. See that testing doesn't work great. Loop through request to gather batch that then gets submitted to the model.
v3: Add in Async for the BLS portion so that we get the highest through put.